+++ chaste_codegen_sbml/SbmlRefModels/src/dynamic/TysonNovak2001/TysonNovak2001SbmlOdeSystem.hpp
@@ -1,24 +1,28 @@
-#ifndef TYSONNOVAK2001SBMLODESYSTEMANDCELLCYCLEMODEL_HPP_
-#define TYSONNOVAK2001SBMLODESYSTEMANDCELLCYCLEMODEL_HPP_
+#ifndef TYSONNOVAK2001SBMLODESYSTEM_HPP_
+#define TYSONNOVAK2001SBMLODESYSTEM_HPP_
 
 #include <vector>
 
 #include <boost/serialization/base_object.hpp>
 
-#include "AbstractOdeSystem.hpp"
+#include "AbstractSbmlOdeSystem.hpp"
 #include "ChasteSerialization.hpp"
 #include "SbmlEventType.hpp"
 
-class TysonNovak2001SbmlOdeSystem : public AbstractOdeSystem
+class TysonNovak2001SbmlOdeSystem : public AbstractSbmlOdeSystem
 {
 private:
-    // (De-)Serialize TysonNovak2001SbmlOdeSystem
     friend class boost::serialization::access;
-
+    /**
+     * Save / load TysonNovak2001SbmlOdeSystem archive
+     *
+     * @param archive the archive
+     * @param version the current version of this class
+     */
     template <class Archive>
-    void serialize(Archive& ar, const unsigned int version)
+    void serialize(Archive& archive, const unsigned int version)
     {
-        ar& BOOST_SERIALIZATION_BASE_OBJECT_NVP(AbstractOdeSystem);
+        archive& BOOST_SERIALIZATION_BASE_OBJECT_NVP(AbstractSbmlOdeSystem);
     }
 
     // CONSTANT PARAMETERS
@@ -102,62 +106,16 @@ private:
     double SKsynthesis;               // SK synthesis
     double SKdegradation;             // SK degradation
 
-    // EVENTS
-    std::vector<bool> mEventSatisfied;
-    std::vector<bool> mEventTriggered;
-    std::vector<SbmlEventType> mEventType;
-    std::vector<bool> mEventAdjustedStateVars;
-    std::vector<double> mEventAdjustedStateValues;
-    std::vector<bool> mEventAdjustedParameters;
-    std::vector<double> mEventAdjustedParameterValues;
-
 public:
     /**
      * Default constructor
-     *
-     * @param stateVariables Initial state variables (optional)
      */
-    TysonNovak2001SbmlOdeSystem(std::vector<double> stateVariables = std::vector<double>());
-
-    /**
-     * Copy constructor
-     *
-     * @param rOdeSystem Reference to the original instance
-     */
-    TysonNovak2001SbmlOdeSystem(const TysonNovak2001SbmlOdeSystem& rOdeSystem);
+    TysonNovak2001SbmlOdeSystem();
 
     /**
      * Destructor
      */
     ~TysonNovak2001SbmlOdeSystem();
-
-    /**
-     * Adjust parameters and state variables after a stopping event
-     *
-     * @param time The current time
-     */
-    void AdjustParameters(double time);
-
-    /**
-     * Calculate whether the conditions to trigger an event have been met
-     * (Used by CVODE solver to find exact stopping position)
-     *
-     * @param time The current time
-     * @param rY The current state variables
-     *
-     * @return How close we are to the root of the stopping condition
-     */
-    double CalculateRootFunction(double time, const std::vector<double>& rY) override;
-
-    /**
-     * Calculate whether the conditions to trigger an event have been met
-     *
-     * @param time The current time
-     * @param rY The current state variables
-     *
-     * @return True if conditions for an event are met, false otherwise
-     */
-    bool CalculateStoppingEvent(double time, const std::vector<double>& rY) override;
 
     /**
      * Compute the derived quantities from the given system state.
@@ -167,7 +125,7 @@ public:
      *
      * @return a vector of derived quantities
      */
-    std::vector<double> ComputeDerivedQuantities(double time, const std::vector<double>& rY);
+    std::vector<double> ComputeDerivedQuantities(double time, const std::vector<double>& rY) override;
 
     /**
      * Compute the RHS of the ODE system.
@@ -181,15 +139,6 @@ public:
     void EvaluateYDerivatives(double time, const std::vector<double>& rY, std::vector<double>& rDY) override;
 
     /**
-     * Check if a specific type of event has occurred.
-     *
-     * @param eventType The type of event to check
-     *
-     * @return True if the type of event has occurred, false otherwise
-     */
-    bool HasEventOccurred(SbmlEventType eventType);
-
-    /**
      * Process the events in the model.
      *
      * @param time The current time
@@ -197,12 +146,7 @@ public:
      *
      * @return How close we are to the time of the next event
      */
-    double ProcessModelEvents(double time, const std::vector<double>& rY);
-
-    /**
-     * Reset the flags that indicate which events have been triggered.
-     */
-    void ResetEventsOccurred();
+    double ProcessModelEvents(double time, const std::vector<double>& rY) override;
 
     /**
      * Run the equations governing the model to update state.
@@ -210,51 +154,14 @@ public:
      * @param time The current time
      * @param rY The current state variables
      */
-    void RunModelRules(double time, const std::vector<double>& rY);
+    void RunModelRules(double time, const std::vector<double>& rY) override;
 
     // MODEL FUNCTIONS
     inline double GK(double A1, double A2, double A3, double A4);
 };
 
-namespace
-{
-namespace serialization
-{
-    // Provide constructor for serializing TysonNovak2001SbmlOdeSystem
-    template <class Archive>
-    inline void save_construct_data(Archive& ar, const TysonNovak2001SbmlOdeSystem* t, const unsigned int version)
-    {
-        // Save data required to construct instance
-        const std::vector<double> state_variables = t->rGetConstStateVariables();
-        ar << state_variables;
-    }
-
-    // Provide constructor for de-serializing TysonNovak2001SbmlOdeSystem
-    template <class Archive>
-    inline void load_construct_data(Archive& ar, TysonNovak2001SbmlOdeSystem* t, const unsigned int version)
-    {
-        // Retrieve data from archive required to construct new instance
-        std::vector<double> state_variables;
-        ar >> state_variables;
-
-        // Invoke inplace constructor to initialise instance
-        ::new (t) TysonNovak2001SbmlOdeSystem(state_variables);
-    }
-} // namespace serialization
-} // namespace
-
-// Define SbmlCellCycleWrapperModel using wrappers
-#include "SbmlCellCycleWrapperModel.cpp"
-#include "SbmlCellCycleWrapperModel.hpp"
-
-typedef SbmlCellCycleWrapperModel<TysonNovak2001SbmlOdeSystem, 8> TysonNovak2001SbmlCellCycleModel;
-
-// Declare identifiers for the serializer
+// Register the ODE system with Boost serialization
 #include "SerializationExportWrapper.hpp"
 CHASTE_CLASS_EXPORT(TysonNovak2001SbmlOdeSystem)
-EXPORT_TEMPLATE_CLASS2(SbmlCellCycleWrapperModel, TysonNovak2001SbmlOdeSystem, 8)
 
-#include "CellCycleModelOdeSolverExportWrapper.hpp"
-EXPORT_CELL_CYCLE_MODEL_ODE_SOLVER(TysonNovak2001SbmlCellCycleModel)
-
-#endif // TYSONNOVAK2001SBMLODESYSTEMANDCELLCYCLEMODEL_HPP_
+#endif // TYSONNOVAK2001SBMLODESYSTEM_HPP_
