+++ chaste_codegen_sbml/SbmlRefModels/test/TestTysonNovak2001SbmlCellCycleModel.hpp
@@ -1,0 +1,213 @@
+/*
+
+Copyright (c) 2005-2025, University of Oxford.
+All rights reserved.
+
+University of Oxford means the Chancellor, Masters and Scholars of the
+University of Oxford, having an administrative office at Wellington
+Square, Oxford OX1 2JD, UK.
+
+This file is part of Chaste.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice,
+   this list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+ * Neither the name of the University of Oxford nor the names of its
+   contributors may be used to endorse or promote products derived from this
+   software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE
+GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
+OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+*/
+
+#ifndef TEST_TYSON_NOVAK_2001_SBML_CELL_CYCLE_MODEL_HPP_
+#define TEST_TYSON_NOVAK_2001_SBML_CELL_CYCLE_MODEL_HPP_
+
+#include <iostream>
+#include <vector>
+
+#include <boost/archive/text_iarchive.hpp>
+#include <boost/archive/text_oarchive.hpp>
+#include <boost/make_shared.hpp>
+#include <boost/shared_ptr.hpp>
+
+#include <cxxtest/TestSuite.h>
+
+#include "AbstractCellBasedTestSuite.hpp"
+#include "ApcOneHitCellMutationState.hpp"
+#include "BackwardEulerIvpOdeSolver.hpp"
+#include "CellCycleModelOdeSolver.hpp"
+#include "StemCellProliferativeType.hpp"
+#include "WildTypeCellMutationState.hpp"
+
+#include "TysonNovak2001SbmlCellCycleModel.hpp"
+
+// This test is never run in parallel
+#include "FakePetscSetup.hpp"
+
+class TestTysonNovak2001SbmlCellCycleModel : public AbstractCellBasedTestSuite
+{
+private:
+    const unsigned ODE_SIZE = 8u;
+
+public:
+    void TestCellCycleModel()
+    {
+        // Setup time
+        SimulationTime* p_simulation_time = SimulationTime::Instance();
+        const unsigned num_timesteps = 10000;
+        p_simulation_time->SetEndTimeAndNumberOfTimeSteps(300.0, num_timesteps);
+
+        // Create a healthy cell
+        auto p_wild_state = boost::make_shared<WildTypeCellMutationState>();
+        auto p_stem_type = boost::make_shared<StemCellProliferativeType>();
+
+        auto p_cell_0 = boost::make_shared<Cell>(p_wild_state, new TysonNovak2001SbmlCellCycleModel);
+        p_cell_0->SetCellProliferativeType(p_stem_type);
+
+        // Set up the cell cycle model - this should use CVODE by default
+        auto p_ccm_0 = static_cast<TysonNovak2001SbmlCellCycleModel*>(p_cell_0->GetCellCycleModel());
+        p_ccm_0->SetBirthTime(p_simulation_time->GetTime());
+        TS_ASSERT_EQUALS(p_ccm_0->CanCellTerminallyDifferentiate(), false);
+
+        p_cell_0->InitialiseCellCycleModel();
+        p_ccm_0->SetDt(0.01);
+
+        // Create another cell with a cell-cycle model that uses a BackwardEulerIvpOdeSolver
+        auto solver = CellCycleModelOdeSolver<TysonNovak2001SbmlCellCycleModel, BackwardEulerIvpOdeSolver>::Instance();
+        boost::shared_ptr<CellCycleModelOdeSolver<TysonNovak2001SbmlCellCycleModel, BackwardEulerIvpOdeSolver> > p_solver(solver);
+        p_solver->SetSizeOfOdeSystem(ODE_SIZE);
+        p_solver->Initialise();
+
+        auto p_cell_1 = boost::make_shared<Cell>(p_wild_state, new TysonNovak2001SbmlCellCycleModel(p_solver));
+        p_cell_1->SetCellProliferativeType(p_stem_type);
+
+        auto p_ccm_1 = static_cast<TysonNovak2001SbmlCellCycleModel*>(p_cell_1->GetCellCycleModel());
+        p_ccm_1->SetBirthTime(p_simulation_time->GetTime());
+        TS_ASSERT_EQUALS(p_ccm_1->CanCellTerminallyDifferentiate(), false);
+        TS_ASSERT_EQUALS(p_ccm_1->GetOdeSolver()->GetSizeOfOdeSystem(), ODE_SIZE);
+
+        p_cell_1->InitialiseCellCycleModel();
+        TS_ASSERT_EQUALS(p_ccm_1->GetDt(), 0.0001); // Timestep for non-adaptive solvers defaults to 0.0001
+        p_ccm_1->SetDt(0.01);
+
+        // Test the cell is ready to divide at the right time
+        double standard_divide_time = 103.80;
+        double tolerance = 0.05;
+        for (unsigned i = 0; i < num_timesteps / 2; i++)
+        {
+            p_simulation_time->IncrementTimeOneStep();
+            double time = p_simulation_time->GetTime();
+
+            bool division_ready_0 = p_ccm_0->ReadyToDivide();
+            bool division_ready_1 = p_ccm_1->ReadyToDivide();
+
+            if (time > standard_divide_time + tolerance)
+            {
+                TS_ASSERT_EQUALS(division_ready_0, true);
+                TS_ASSERT_EQUALS(division_ready_1, true);
+            }
+            else if (time < standard_divide_time - tolerance)
+            {
+                TS_ASSERT_EQUALS(division_ready_0, false);
+                TS_ASSERT_EQUALS(division_ready_1, false);
+            }
+        }
+
+        // Check CVODE vs BackwardEuler solution
+        std::vector<double> proteins_0 = p_ccm_0->GetProteinConcentrations();
+        TS_ASSERT_EQUALS(proteins_0.size(), ODE_SIZE);
+        TS_ASSERT_DELTA(proteins_0[0], 0.1789, 1e-4); // CycBt
+        TS_ASSERT_DELTA(proteins_0[1], 0.3039, 1e-4); // Cdc20a
+        TS_ASSERT_DELTA(proteins_0[2], 0.4455, 1e-4); // Cdh1
+        TS_ASSERT_DELTA(proteins_0[3], 0.8125, 1e-4); // m
+        TS_ASSERT_DELTA(proteins_0[4], 1.1626, 1e-4); // Cdc20t
+        TS_ASSERT_DELTA(proteins_0[5], 0.5465, 1e-4); // IEP
+        TS_ASSERT_DELTA(proteins_0[6], 0.0800, 1e-4); // CKIt
+        TS_ASSERT_DELTA(proteins_0[7], 0.0816, 1e-4); // SK
+
+        std::vector<double> proteins_1 = p_ccm_1->GetProteinConcentrations();
+        TS_ASSERT_EQUALS(proteins_1.size(), ODE_SIZE);
+        TS_ASSERT_DELTA(proteins_1[0], proteins_0[0], 1e-2); // CycBt
+        TS_ASSERT_DELTA(proteins_1[1], proteins_0[1], 1e-2); // Cdc20a
+        TS_ASSERT_DELTA(proteins_1[2], proteins_0[2], 1e-2); // Cdh1
+        TS_ASSERT_DELTA(proteins_1[3], proteins_0[3], 1e-2); // m
+        TS_ASSERT_DELTA(proteins_1[4], proteins_0[4], 1e-2); // Cdc20t
+        TS_ASSERT_DELTA(proteins_1[5], proteins_0[5], 1e-2); // IEP
+        TS_ASSERT_DELTA(proteins_1[6], proteins_0[6], 1e-2); // CKIt
+        TS_ASSERT_DELTA(proteins_1[7], proteins_0[7], 1e-2); // SK
+
+        // Test for a mutant cell
+        TS_ASSERT_EQUALS(p_ccm_0->ReadyToDivide(), true);
+        p_ccm_0->ResetForDivision();
+        TS_ASSERT_EQUALS(p_ccm_0->ReadyToDivide(), false);
+
+        auto p_mutation = boost::make_shared<ApcOneHitCellMutationState>();
+        auto p_ccm_2 = static_cast<TysonNovak2001SbmlCellCycleModel*>(p_ccm_0->CreateCellCycleModel());
+        auto p_cell_2 = boost::make_shared<Cell>(p_mutation, p_ccm_2);
+        p_cell_2->SetCellProliferativeType(p_stem_type);
+
+        TS_ASSERT_EQUALS(p_cell_2->ReadyToDivide(), false);
+        TS_ASSERT_EQUALS(p_ccm_2->ReadyToDivide(), false);
+
+        // Test the cell is ready to divide at the right time
+        standard_divide_time = 250.32;
+        for (unsigned i = 0; i < num_timesteps / 2; i++)
+        {
+            p_simulation_time->IncrementTimeOneStep();
+            double time = p_simulation_time->GetTime();
+
+            bool division_ready_0 = p_ccm_0->ReadyToDivide();
+            bool division_ready_2 = p_ccm_2->ReadyToDivide();
+
+            if (time > standard_divide_time)
+            {
+                TS_ASSERT_EQUALS(division_ready_0, true);
+                TS_ASSERT_EQUALS(division_ready_2, true);
+            }
+            else if (time < standard_divide_time)
+            {
+                TS_ASSERT_EQUALS(division_ready_0, false);
+                TS_ASSERT_EQUALS(division_ready_2, false);
+            }
+        }
+
+        // Check ODE solution
+        proteins_0 = p_ccm_0->GetProteinConcentrations();
+        TS_ASSERT_EQUALS(proteins_0.size(), ODE_SIZE);
+        TS_ASSERT_DELTA(proteins_0[0], 0.1794, 1e-4); // CycBt
+        TS_ASSERT_DELTA(proteins_0[1], 0.2977, 1e-4); // Cdc20a
+        TS_ASSERT_DELTA(proteins_0[2], 0.4349, 1e-4); // Cdh1
+        TS_ASSERT_DELTA(proteins_0[3], 0.8096, 1e-4); // m
+        TS_ASSERT_DELTA(proteins_0[4], 1.1307, 1e-4); // Cdc20t
+        TS_ASSERT_DELTA(proteins_0[5], 0.5447, 1e-4); // IEP
+        TS_ASSERT_DELTA(proteins_0[6], 0.0806, 1e-4); // CKIt
+        TS_ASSERT_DELTA(proteins_0[7], 0.0799, 1e-4); // SK
+
+        std::vector<double> proteins_2 = p_ccm_2->GetProteinConcentrations();
+        TS_ASSERT_EQUALS(proteins_2.size(), ODE_SIZE);
+        TS_ASSERT_DELTA(proteins_2[0], proteins_0[0], 1e-4); // CycBt
+        TS_ASSERT_DELTA(proteins_2[1], proteins_0[1], 1e-4); // Cdc20a
+        TS_ASSERT_DELTA(proteins_2[2], proteins_0[2], 1e-4); // Cdh1
+        TS_ASSERT_DELTA(proteins_2[3], proteins_0[3], 1e-4); // m
+        TS_ASSERT_DELTA(proteins_2[4], proteins_0[4], 1e-4); // Cdc20t
+        TS_ASSERT_DELTA(proteins_2[5], proteins_0[5], 1e-4); // IEP
+        TS_ASSERT_DELTA(proteins_2[6], proteins_0[6], 1e-4); // CKIt
+        TS_ASSERT_DELTA(proteins_2[7], proteins_0[7], 1e-4); // SK
+    }
+};
+
+#endif // TEST_TYSON_NOVAK_2001_SBML_CELL_CYCLE_MODEL_HPP_
